refactor: Made minor improvements to Controllers
