+++ src/Controller/Data/Hospital/HospitalShowController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller\Data;
+namespace App\Controller\Data\Hospital;
 
 use App\Entity\Hospital;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -9,7 +9,7 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class HospitalShowController extends AbstractController
 {
-    #[Route('/hospital/show/{id}', name: 'app_data_hospital_show')]
+    #[Route('/data/hospital/{id}', name: 'app_data_hospital_show')]
     public function index(Hospital $hospital): Response
     {
         return $this->render('data/hospital/show.html.twig', [
